Move functions into TableFilter
